refactor: update Message component to use theme styles and fix theme colors
